Video user querys

+++ src/db/querys/article_snippets_query.rs
@@ -18,9 +18,9 @@ fn row_to_article_snippet(row: rocket_contrib::databases::postgres::rows::Row) -
     }
 }
 
-pub fn delete_article_snippet(conn: &TigumPgConn, id: i32) -> Json<String> {
+pub fn delete_article_snippet(conn: &TigumPgConn, id: i32, user_id: i32) -> Json<String> {
     let update = conn
-        .execute("DELETE FROM article_snippets WHERE id = $1", &[&id])
+        .execute("DELETE FROM article_snippets WHERE id = $1 AND user_id = $2", &[&id, &user_id])
         .unwrap();
     Json(format!("{} rows affected", update))
 }
@@ -29,10 +29,11 @@ pub fn update_article_snippet(
     conn: &TigumPgConn,
     id: i32,
     article_snippet: Json<NewArticleSnippet>,
+    user_id: i32
 ) -> Json<ArticleSnippet> {
     let updated_rows = conn.query(
-        "UPDATE article_snippets SET topic_id = $2, user_id = $3, content = $4, origin = $5 WHERE id = $1 RETURNING *",
-        &[&id, &article_snippet.topic_id, &article_snippet.user_id, &article_snippet.content, &article_snippet.origin],
+        "UPDATE article_snippets SET topic_id = $2, user_id = $3, content = $4, origin = $5 WHERE id = $1 AND user_id = $3 RETURNING *",
+        &[&id, &article_snippet.topic_id, &user_id, &article_snippet.content, &article_snippet.origin],
     ).unwrap();
 
     let article_snippet_response = row_to_article_snippet(updated_rows.get(0));
@@ -40,12 +41,12 @@ pub fn update_article_snippet(
     Json(article_snippet_response)
 }
 
-pub fn get_article_snippets(conn: &TigumPgConn, ids: Json<Ids>) -> Json<Vec<ArticleSnippet>> {
+pub fn get_article_snippets(conn: &TigumPgConn, ids: Json<Ids>, user_id: i32) -> Json<Vec<ArticleSnippet>> {
     println!("{:?}", ids);
     let query_result = conn
         .query(
-            "SELECT * FROM article_snippets WHERE id = ANY($1)",
-            &[&ids.ids],
+            "SELECT * FROM article_snippets WHERE id = ANY($1) AND user_id = $2",
+            &[&ids.ids, &user_id],
         )
         .unwrap();
     let mut results: Vec<ArticleSnippet> = vec![];
@@ -56,9 +57,9 @@ pub fn get_article_snippets(conn: &TigumPgConn, ids: Json<Ids>) -> Json<Vec<Arti
     Json(results)
 }
 
-pub fn get_article_snippet(conn: &TigumPgConn, id: i32) -> Json<ArticleSnippet> {
+pub fn get_article_snippet(conn: &TigumPgConn, id: i32, user_id: i32) -> Json<ArticleSnippet> {
     let query_result = conn
-        .query("SELECT * FROM article_snippets WHERE id = $1", &[&id])
+        .query("SELECT * FROM article_snippets WHERE id = $1 AND user_id = $2", &[&id, &user_id])
         .unwrap();
     println!("{:#?}", query_result);
     let article_snippet_response = row_to_article_snippet(query_result.get(0));
@@ -68,6 +69,7 @@ pub fn get_article_snippet(conn: &TigumPgConn, id: i32) -> Json<ArticleSnippet> 
 pub fn create_article_snippet(
     conn: &TigumPgConn,
     article_snippet: &Json<NewArticleSnippet>,
+    user_id: i32
 ) -> Json<ArticleSnippet> {
     let inserted_row = conn
         .query(
@@ -76,7 +78,7 @@ pub fn create_article_snippet(
                 &article_snippet.content,
                 &article_snippet.origin,
                 &article_snippet.topic_id,
-                &article_snippet.user_id,
+                &user_id,
             ],
         )
         .unwrap();
