+++ src/db/querys/video_query.rs
@@ -20,17 +20,17 @@ fn row_to_video(row: rocket_contrib::databases::postgres::rows::Row) -> Video {
     }
 }
 
-pub fn delete_video(conn: &TigumPgConn, id: i32) -> Json<String> {
+pub fn delete_video(conn: &TigumPgConn, id: i32, user_id: i32) -> Json<String> {
     let update = conn
-        .execute("DELETE FROM videos WHERE id = $1", &[&id])
+        .execute("DELETE FROM videos WHERE id = $1 AND user_id = $2", &[&id, &user_id])
         .unwrap();
     Json(format!("{} rows affected", update))
 }
 
-pub fn update_video(conn: &TigumPgConn, id: i32, video: Json<NewVideo>) -> Json<Video> {
+pub fn update_video(conn: &TigumPgConn, id: i32, video: Json<NewVideo>, user_id: i32) -> Json<Video> {
     let updated_rows = conn.query(
         "UPDATE videos SET topic_id = $2, user_id = $3, title = $4, iframe = $5, origin = $6, thumbnail_img = $7 WHERE id = $1 RETURNING *",
-        &[&id, &video.topic_id, &video.user_id, &video.title, &video.iframe, &video.origin, &video.thumbnail_img],
+        &[&id, &video.topic_id, &user_id, &video.title, &video.iframe, &video.origin, &video.thumbnail_img],
     ).unwrap();
 
     let video_response = row_to_video(updated_rows.get(0));
@@ -38,10 +38,9 @@ pub fn update_video(conn: &TigumPgConn, id: i32, video: Json<NewVideo>) -> Json<
     Json(video_response)
 }
 
-pub fn get_videos(conn: &TigumPgConn, ids: Json<Ids>) -> Json<Vec<Video>> {
-    println!("{:?}", ids);
+pub fn get_videos(conn: &TigumPgConn, ids: Json<Ids>, user_id: i32) -> Json<Vec<Video>> {
     let query_result = conn
-        .query("SELECT * FROM videos WHERE id = ANY($1)", &[&ids.ids])
+        .query("SELECT * FROM videos WHERE id = ANY($1) AND user_id = $2", &[&ids.ids, &user_id])
         .unwrap();
     let mut results: Vec<Video> = vec![];
     for row in query_result.iter() {
@@ -51,7 +50,7 @@ pub fn get_videos(conn: &TigumPgConn, ids: Json<Ids>) -> Json<Vec<Video>> {
     Json(results)
 }
 
-pub fn get_video(conn: &TigumPgConn, id: i32) -> Json<Video> {
+pub fn get_video(conn: &TigumPgConn, id: i32, user_id: i32) -> Json<Video> {
     let query_result = conn
         .query("SELECT * FROM videos WHERE id = $1", &[&id])
         .unwrap();
@@ -59,13 +58,13 @@ pub fn get_video(conn: &TigumPgConn, id: i32) -> Json<Video> {
     Json(video_response)
 }
 
-pub fn create_video(conn: &TigumPgConn, video: &Json<NewVideo>) -> Json<Video> {
+pub fn create_video(conn: &TigumPgConn, video: &Json<NewVideo>, user_id: i32) -> Json<Video> {
     let inserted_row = conn
         .query(
             "INSERT INTO videos (topic_id, user_id, title, iframe, origin, thumbnail_img) VALUES ($1, $2, $3, $4, $5, $6) RETURNING *",
             &[
                 &video.topic_id,
-                &video.user_id,
+                &user_id,
                 &video.title,
                 &video.iframe,
                 &video.origin,
